+++ src/base/browser/secondary/listView/listCache.ts
@@ -87,7 +87,7 @@ export class ListViewCache implements IDisposable {
      * @param row The row about to be released.
      */
     public release(row: IListViewRow): void {
-        DomUtility.Elements.removeNodeFromParent(row.dom);
+        DomUtility.Modifiers.removeNodeFromParent(row.dom);
         const cache = this.__getCache(row.type);
         cache.push(row);
     }
